vrt creation in rastercalculator
used when bounds of inputrasters are not the same
