+++ tests_hrt/test_raster_functions.py
@@ -64,38 +64,6 @@ class TestRasterFunctions:
         assert vrt_path.exists()
 
 
-def test_raster_calculator():
-    raster_depth = hrt.Raster(TEST_DIRECTORY / r"depth_test.tif")
-    raster_lu = hrt.Raster(TEST_DIRECTORY / r"landuse_test.tif")
-    raster_out = hrt.Raster(TEMP_DIR / f"rastercalc_{hrt.get_uuid()}.tif")
-
-    def run_window(block):
-        block_out = block.blocks["lu"]
-
-        # Nodatamasks toepassen
-        block_out[block.masks_all] = 0
-        return block_out
-
-    calc = hrt.RasterCalculatorV2(
-        raster_out=raster_out,
-        raster_paths_dict={
-            "depth": raster_depth,
-            "lu": raster_lu,
-        },
-        nodata_keys=["lu"],
-        mask_keys=["depth", "lu"],
-        metadata_key="depth",
-        custom_run_window_function=run_window,
-        output_nodata=0,
-        min_block_size=4096,
-        verbose=True,
-    )
-
-    calc.run(overwrite=False)
-
-    assert raster_out.sum() == 261581
-
-
 if __name__ == "__main__":
     import inspect
 
@@ -111,6 +79,5 @@ if __name__ == "__main__":
     #         print(i)
     #         getattr(selftest, i)()
 
-    test_raster_calculator()
 
 # %%
